updates to settng page
