+++ interface/src/framework/system/UploadDownload.tsx
@@ -1,12 +1,12 @@
 import DownloadIcon from '@mui/icons-material/GetApp';
-import { Typography, Button, Box } from '@mui/material';
+import { Typography, Button, Box, Link } from '@mui/material';
 import { useRequest } from 'alova';
 import { useState, type FC } from 'react';
 import { toast } from 'react-toastify';
 import RestartMonitor from './RestartMonitor';
 
 import * as SystemApi from 'api/system';
-import { SectionContent, SingleUpload, useLayoutTitle } from 'components';
+import { FormLoader, SectionContent, SingleUpload, useLayoutTitle } from 'components';
 
 import { useI18nContext } from 'i18n/i18n-react';
 import * as EMSESP from 'project/api';
@@ -31,6 +31,26 @@ const UploadDownload: FC = () => {
   const { send: getAPI, onSuccess: onGetAPI } = useRequest((data) => EMSESP.API(data), {
     immediate: false
   });
+
+  const { data: data, send: loadData, error } = useRequest(SystemApi.readSystemStatus, { force: true });
+
+  const { data: latestVersion } = useRequest(SystemApi.getStableVersion, {
+    immediate: true,
+    force: true
+  });
+  const { data: latestDevVersion } = useRequest(SystemApi.getDevVersion, {
+    immediate: true,
+    force: true
+  });
+
+  const STABLE_URL = 'https://github.com/emsesp/EMS-ESP32/releases/download/';
+  const DEV_URL = 'https://github.com/emsesp/EMS-ESP32/releases/download/latest/';
+
+  const STABLE_RELNOTES_URL = 'https://github.com/emsesp/EMS-ESP32/blob/main/CHANGELOG.md';
+  const DEV_RELNOTES_URL = 'https://github.com/emsesp/EMS-ESP32/blob/dev/CHANGELOG_LATEST.md';
+
+  const getBinURL = (v: string) =>
+    'EMS-ESP-' + v.replaceAll('.', '_') + '-' + data.esp_platform.replaceAll('-', '_') + '.bin';
 
   const {
     loading: isUploading,
@@ -129,88 +149,158 @@ const UploadDownload: FC = () => {
 
   useLayoutTitle(LL.UPLOAD_DOWNLOAD());
 
-  const content = () => (
-    <>
-      <Typography sx={{ pt: 2, pb: 2 }} variant="h6" color="primary">
-        {LL.UPLOAD()}
-      </Typography>
-      <Box mb={2} color="warning.main">
-        <Typography variant="body2">
-          {LL.UPLOAD_TEXT()}
-          <br />
-          <br />
-          {LL.RESTART_TEXT(1)}.
-        </Typography>
-      </Box>
-      {md5 && (
-        <Box mb={2}>
-          <Typography variant="body2">{'MD5: ' + md5}</Typography>
-        </Box>
-      )}
-      <SingleUpload onDrop={startUpload} onCancel={cancelUpload} isUploading={isUploading} progress={progress} />
-      {!isUploading && (
-        <>
-          <Typography sx={{ pt: 4, pb: 2 }} variant="h6" color="primary">
-            {LL.DOWNLOAD(0)}&nbsp;{LL.SUPPORT_INFORMATION(1)}
-          </Typography>
-          <Button
-            startIcon={<DownloadIcon />}
-            variant="outlined"
-            color="primary"
-            onClick={() => callAPI('system', 'info')}
-          >
-            {LL.SUPPORT_INFORMATION(0)}
-          </Button>
-          <Button
-            sx={{ ml: 2 }}
-            startIcon={<DownloadIcon />}
-            variant="outlined"
-            color="primary"
-            onClick={() => callAPI('system', 'allvalues')}
-          >
-            All Values
-          </Button>
+  const content = () => {
+    if (!data) {
+      return <FormLoader onRetry={loadData} errorMessage={error?.message} />;
+    }
 
-          <Typography sx={{ pt: 4, pb: 2 }} variant="h6" color="primary">
-            {LL.DOWNLOAD(0)}&nbsp;{LL.SETTINGS(1)}
+    return (
+      <>
+        <Typography sx={{ pb: 2 }} variant="h6" color="primary">
+          {LL.EMS_ESP_VER()}
+        </Typography>
+        <Box p={2} border="2px solid grey" borderRadius={2}>
+          {LL.VERSION_ON() + ' '}
+          <b>{data.emsesp_version}</b>&nbsp;({data.esp_platform})
+          {latestVersion && (
+            <Box mt={2}>
+              {LL.THE_LATEST()}&nbsp;{LL.OFFICIAL()}&nbsp;{LL.RELEASE_IS()}&nbsp;<b>{latestVersion}</b>
+              &nbsp;(
+              <Link target="_blank" href={STABLE_RELNOTES_URL} color="primary">
+                {LL.RELEASE_NOTES()}
+              </Link>
+              )&nbsp;(
+              <Link
+                target="_blank"
+                href={STABLE_URL + 'v' + latestVersion + '/' + getBinURL(latestVersion)}
+                color="primary"
+              >
+                {LL.DOWNLOAD(1)}
+              </Link>
+              )
+            </Box>
+          )}
+          {latestDevVersion && (
+            <Box mt={2}>
+              {LL.THE_LATEST()}&nbsp;{LL.DEVELOPMENT()}&nbsp;{LL.RELEASE_IS()}&nbsp;
+              <b>{latestDevVersion}</b>
+              &nbsp;(
+              <Link target="_blank" href={DEV_RELNOTES_URL} color="primary">
+                {LL.RELEASE_NOTES()}
+              </Link>
+              )&nbsp;(
+              <Link target="_blank" href={DEV_URL + getBinURL(latestDevVersion)} color="primary">
+                {LL.DOWNLOAD(1)}
+              </Link>
+              )
+            </Box>
+          )}
+        </Box>
+
+        <Typography sx={{ pt: 2, pb: 2 }} variant="h6" color="primary">
+          {LL.UPLOAD()}
+        </Typography>
+        <Box mb={2} color="warning.main">
+          <Typography variant="body2">
+            {LL.UPLOAD_TEXT()}
+            <br />
+            <br />
+            {LL.RESTART_TEXT(1)}.
           </Typography>
-          <Box color="warning.main">
-            <Typography mb={1} variant="body2">
-              {LL.DOWNLOAD_SETTINGS_TEXT()}
-            </Typography>
+        </Box>
+        {md5 && (
+          <Box mb={2}>
+            <Typography variant="body2">{'MD5: ' + md5}</Typography>
           </Box>
-          <Button startIcon={<DownloadIcon />} variant="outlined" color="primary" onClick={downloadSettings}>
-            {LL.SETTINGS_OF('')}
-          </Button>
-          <Box color="warning.main">
-            <Typography mt={2} mb={1} variant="body2">
-              {LL.DOWNLOAD_CUSTOMIZATION_TEXT()}
+        )}
+        <SingleUpload onDrop={startUpload} onCancel={cancelUpload} isUploading={isUploading} progress={progress} />
+        {!isUploading && (
+          <>
+            <Typography sx={{ pt: 4, pb: 2 }} variant="h6" color="primary">
+              {LL.DOWNLOAD(0)}
             </Typography>
-          </Box>
-          <Button startIcon={<DownloadIcon />} variant="outlined" color="primary" onClick={downloadCustomizations}>
-            {LL.CUSTOMIZATIONS()}
-          </Button>
-          <Button
-            sx={{ ml: 2 }}
-            startIcon={<DownloadIcon />}
-            variant="outlined"
-            color="primary"
-            onClick={downloadEntities}
-          >
-            {LL.CUSTOM_ENTITIES(0)}
-          </Button>
-          <Box color="warning.main">
-            <Typography mt={2} mb={1} variant="body2">
-              {LL.DOWNLOAD_SCHEDULE_TEXT()}
-            </Typography>
-          </Box>
-          <Button startIcon={<DownloadIcon />} variant="outlined" color="primary" onClick={downloadSchedule}>
-            {LL.SCHEDULE(0)}
-          </Button>
-        </>
-      )}
-    </>
-  );
+            <Box color="warning.main">
+              <Typography mb={1} variant="body2">
+                {LL.HELP_INFORMATION_4()}
+              </Typography>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={() => callAPI('system', 'info')}
+              >
+                {LL.SUPPORT_INFORMATION(0)}
+              </Button>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={() => callAPI('system', 'allvalues')}
+              >
+                All Values
+              </Button>
+            </Box>
+
+            <Box color="warning.main">
+              <Typography mt={2} mb={1} variant="body2">
+                {LL.DOWNLOAD_SETTINGS_TEXT()}
+              </Typography>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={downloadSettings}
+              >
+                {LL.SETTINGS_OF('')}
+              </Button>
+            </Box>
+
+            <Box color="warning.main">
+              <Typography mt={2} mb={1} variant="body2">
+                {LL.DOWNLOAD_CUSTOMIZATION_TEXT()}
+              </Typography>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={downloadCustomizations}
+              >
+                {LL.CUSTOMIZATIONS()}
+              </Button>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={downloadEntities}
+              >
+                {LL.CUSTOM_ENTITIES(0)}
+              </Button>
+              <Box color="warning.main">
+                <Typography mt={2} mb={1} variant="body2">
+                  {LL.DOWNLOAD_SCHEDULE_TEXT()}
+                </Typography>
+              </Box>
+              <Button
+                sx={{ ml: 2 }}
+                startIcon={<DownloadIcon />}
+                variant="outlined"
+                color="primary"
+                onClick={downloadSchedule}
+              >
+                {LL.SCHEDULE(0)}
+              </Button>
+            </Box>
+          </>
+        )}
+      </>
+    );
+  };
+
   return <SectionContent>{restarting ? <RestartMonitor /> : content()}</SectionContent>;
 };
