+++ interface/src/i18n/en/index.ts
@@ -176,13 +176,11 @@ const en: Translation = {
   STATUS_OF: '{0} Status',
   UPLOAD_DOWNLOAD: 'Upload/Download',
   VERSION_ON: 'You are currently on version',
-  SYSTEM_APPLY_FIRMWARE: 'to apply the new firmware',
   CLOSE: 'Close',
   USE: 'Use',
   FACTORY_RESET: 'Factory Reset',
   SYSTEM_FACTORY_TEXT: 'Device has been factory reset and will now restart',
   SYSTEM_FACTORY_TEXT_DIALOG: 'Are you sure you want to reset EMS-ESP to its factory defaults?',
-  VERSION_CHECK: 'Version Check',
   THE_LATEST: 'The latest',
   OFFICIAL: 'official',
   DEVELOPMENT: 'development',
