+++ interface/src/i18n/it/index.ts
@@ -178,13 +178,11 @@ const it: Translation = {
   STATUS_OF: 'Stato {0}',
   UPLOAD_DOWNLOAD: 'Caricamento/Scaricamento',
   VERSION_ON: 'Attualmente stai eseguendo la versione',
-  SYSTEM_APPLY_FIRMWARE: 'per applicare il nuovo firmware',
   CLOSE: 'Chiudere',
   USE: 'Usa',
   FACTORY_RESET: 'Impostazioni di fabbrica',
   SYSTEM_FACTORY_TEXT: 'Il dispositivo è stato ripristinato alle impostazioni di fabbrica e ora verrà riavviato',
   SYSTEM_FACTORY_TEXT_DIALOG: 'Sei sicuro di voler ripristinare il dispositivo alle impostazioni di fabbrica??',
-  VERSION_CHECK: 'Verifica Versione',
   THE_LATEST: 'Ultima',
   OFFICIAL: 'ufficiale',
   DEVELOPMENT: 'sviluppo',
