+++ interface/src/i18n/sk/index.ts
@@ -176,13 +176,11 @@ const sk: Translation = {
   STATUS_OF: '{0} Stav',
   UPLOAD_DOWNLOAD: 'Nahrať/Stiahnuť',
   VERSION_ON: 'Momentálne ste vo verzii',
-  SYSTEM_APPLY_FIRMWARE: 'na použitie nového firmvéru',
   CLOSE: 'Zatvoriť',
   USE: 'Použiť',
   FACTORY_RESET: 'Továrenské nastavenia',
   SYSTEM_FACTORY_TEXT: 'Zariadenie bolo obnovené z výroby a teraz sa reštartuje',
   SYSTEM_FACTORY_TEXT_DIALOG: 'Naozaj chcete resetovať EMS-ESP na predvolené výrobné nastavenia?',
-  VERSION_CHECK: 'Kontrola verzie',
   THE_LATEST: 'Posledná',
   OFFICIAL: 'officiálna',
   DEVELOPMENT: 'vývojárska',
